# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Time Management Project",
    "description": "Website for tracking the working hours of employees of an IT company",
    "version": "1.0.0"
  },
  "host": "localhost:9000",
...
  "schemes": "http",
  "paths": {
...
    "/tasks/get/{id_project}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Returns a tasks by project id.",
        "description": "Endpoint to get a tasks by project id.",
        "parameters": [
          {
            "name": "id_project",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "title",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "developer_id",
            "in": "query",
            "required": true,
            "type": "integer"
          },
           {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tasks/get_single/{id_task}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Returns a task by id.",
        "description": "Endpoint to get a task by  id.",
        "parameters": [
          {
            "name": "id_task",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tasks/count/{id_project}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Returns a count by project id.",
        "description": "Endpoint to get a count by project id.",
        "parameters": [
          {
            "name": "id_project",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "title",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "developer_id",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tasks/delete/{id_task}": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Delete a task.",
        "description": "Endpoint to delete a task.",
        "parameters": [
          {
            "name": "id_task",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/tasks/add_developer": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Add developer to task.",
        "description": "Endpoint to add developer to task. developerId: integer, taskId: integer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tasks/put": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Put task.",
        "description": "Endpoint to put task. title: string, description: string, results: string, actualDeadline: date-time, plannedDeadline: date-time, leadId: integer, projectId integer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tasks/get_by_developer/{developer_id}": {
      "get": {
        "tags": ["Tasks"],

        "summary": "Returns a tasks by developer id.",
        "description": "Endpoint to get a tasks by developer id.",
        "parameters": [
          {
            "name": "developerId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
                 }
                     }
              }
                },
    "/tasks/update/{taskId}": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Update task.",
        "description": "Endpoint update to task. title: string, description: string, results: string, leadId: integer",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}